+++ core/src/AdaptivePackedCellContainer/ConstructorsOfAPC/FabricToAPCLinker.cpp
@@ -7,7 +7,7 @@ namespace PredictedAdaptedEncoding
 
     bool FabricToAPCLinker::BindExternalRawFabricBacking_(
         uint64_t* words_raw,
-        uint16_t cell_count,
+        uint32_t cell_count,
         VagueTemoraryPremativeFabric* fabric_owner,
         uint64_t fabric_slot_idx
     ) noexcept
@@ -53,8 +53,8 @@ namespace PredictedAdaptedEncoding
     }
 
     bool FabricToAPCLinker::CompareExchangeSequentiallRevertInFail(
-        uint16_t starting_idx_in_apc,
-        uint16_t sequential_number_of_cells,
+        uint32_t starting_idx_in_apc,
+        uint8_t sequential_number_of_cells,
         const uint64_t* source_cells
     ) noexcept
     {
